Implemented the directory test Fixed bugs shown by the tests

+++ src/main/java/net/bplaced/clayn/cfs/impl/local/CFSSimpleFileImpl.java
@@ -49,6 +49,8 @@ public class CFSSimpleFileImpl implements SimpleFile
     @Override
     public void delete() throws IOException
     {
+        if(!exists())
+            return;
         Files.delete(realFile);
     }
